Make graphql method more flexible
There are two types or requests

- Queries
- Mutations

These both support variables and aren't always valid JSON.

For example

```graphql
query($number_of_repos:Int!) {
  viewer {
    name
     repositories(last: $number_of_repos) {
       nodes {
         name
       }
     }
   }
}
variables {
   "number_of_repos": 3
}
```

In the original method this would not be possible because we assume
there won't ever be variables.

This will rely on the user making the entire query/mutation correctly.